tech-adapter/pyproject.toml
[tool.poetry]
name = "mongodb-tech-adapter"
version = "0.1.0"
description = ""
authors = ["Agile Lab"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "~3.11"
virtualenv = "^20.26.6"
pytest = "^8.2.1"
pytest-asyncio = "^0.23.7"
mypy = "^1.1.1"
fastapi = {version = "^0.116.1", extras = ["all"]}
pre-commit = "^3.1.1"
ruff = "^0.4.5"
opentelemetry-distro = "^0.55b1"
opentelemetry-instrumentation-fastapi = "^0.55b1"
opentelemetry-exporter-otlp-proto-grpc = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"
pip-audit = "^2.5.3"
pytest-cov = "^5.0.0"
pyyaml = "^6.0"
types-pyyaml = "^6.0"
types-requests = "^2.32.4"
urllib3 = "^2.5.0"
loguru = "^0.7.3"
httpx = "^0.28.0"
requests = "^2.32.4"
jinja2 = "^3.1.6"
h11 = "^0.16.0"
pymongo = "4.13.2"
boolean-py = "^4.0"
starlette = "0.47.2"
pydantic-settings = "^2.10.1"

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
addopts = "-v"

[tool.coverage.report]
fail_under = 90

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.liccheck]
authorized_licenses = [
    'apache license 2.0',
    'apache-2.0',
    'Apache 2.0',
    'apache software',
    'bsd',
    'cc0 1.0 universal (cc0 1.0) public domain dedication',
    'gnu lesser general public license v2 or later (lgplv2+)',
    'isc',
    'isc license (iscl)',
    'mit',
    'mozilla public license 2.0 (mpl 2.0)',
    'other/proprietary',
    'python software foundation',
    'PSF-2.0',
    'the unlicense (unlicense)',
    'mit',
    'MIT License',
    'mozilla public license 2.0 (mpl 2.0)',
    'other/proprietary',
    'python software foundation',
    'the unlicense (unlicense)',
    '3-clause bsd',
    'BSD-2-Clause',
    'BSD-3-Clause',
]
unauthorized_licenses = [
    "gnu general public license v2 (gplv2)"
]


# pytest
[pytest]
asyncio_mode = "auto"
